+++ src/main/java/com/CuteNekoDragon/Core/common/datagen/GlobalDatagenArrays.java
@@ -2,6 +2,7 @@ package com.CuteNekoDragon.Core.common.datagen;
 
 import net.minecraft.world.item.Item;
 import net.p3pp3rf1y.sophisticatedbackpacks.init.ModItems;
+import net.p3pp3rf1y.sophisticatedstorage.init.ModBlocks;
 
 import java.lang.reflect.Array;
 import java.util.ArrayList;
@@ -27,6 +28,38 @@ public class GlobalDatagenArrays {
             ModItems.ADVANCED_TOOL_SWAPPER_UPGRADE.get(),
             net.p3pp3rf1y.sophisticatedstorage.init.ModItems.FEEDING_UPGRADE.get(),
             net.p3pp3rf1y.sophisticatedstorage.init.ModItems.ADVANCED_FEEDING_UPGRADE.get(),
-            net.p3pp3rf1y.sophisticatedstorage.init.ModItems.STACK_UPGRADE_TIER_1_PLUS.get()
+            net.p3pp3rf1y.sophisticatedstorage.init.ModItems.STACK_UPGRADE_TIER_1_PLUS.get(),
+            net.p3pp3rf1y.sophisticatedstorage.init.ModItems.STACK_UPGRADE_TIER_1_TO_TIER_1_PLUS_CONVERSION.get(),
+            net.p3pp3rf1y.sophisticatedstorage.init.ModItems.STACK_UPGRADE_TIER_1_TO_TIER_2_CONVERSION.get(),
+            net.p3pp3rf1y.sophisticatedstorage.init.ModItems.STACK_UPGRADE_TIER_1_TO_TIER_3_CONVERSION.get(),
+            net.p3pp3rf1y.sophisticatedstorage.init.ModItems.STACK_UPGRADE_TIER_1_TO_TIER_4_CONVERSION.get(),
+            net.p3pp3rf1y.sophisticatedstorage.init.ModItems.STACK_UPGRADE_TIER_1_TO_TIER_5_CONVERSION.get(),
+            net.p3pp3rf1y.sophisticatedstorage.init.ModItems.STACK_UPGRADE_TIER_1_PLUS_TO_TIER_2_CONVERSION.get(),
+            net.p3pp3rf1y.sophisticatedstorage.init.ModItems.STACK_UPGRADE_TIER_1_PLUS_TO_TIER_3_CONVERSION.get(),
+            net.p3pp3rf1y.sophisticatedstorage.init.ModItems.STACK_UPGRADE_TIER_1_PLUS_TO_TIER_4_CONVERSION.get(),
+            net.p3pp3rf1y.sophisticatedstorage.init.ModItems.STACK_UPGRADE_TIER_1_PLUS_TO_TIER_5_CONVERSION.get(),
+            net.p3pp3rf1y.sophisticatedstorage.init.ModItems.STACK_UPGRADE_TIER_2_TO_TIER_3_CONVERSION.get(),
+            net.p3pp3rf1y.sophisticatedstorage.init.ModItems.STACK_UPGRADE_TIER_2_TO_TIER_4_CONVERSION.get(),
+            net.p3pp3rf1y.sophisticatedstorage.init.ModItems.STACK_UPGRADE_TIER_2_TO_TIER_5_CONVERSION.get(),
+            net.p3pp3rf1y.sophisticatedstorage.init.ModItems.STACK_UPGRADE_TIER_3_TO_TIER_4_CONVERSION.get(),
+            net.p3pp3rf1y.sophisticatedstorage.init.ModItems.STACK_UPGRADE_TIER_3_TO_TIER_5_CONVERSION.get(),
+            net.p3pp3rf1y.sophisticatedstorage.init.ModItems.STACK_UPGRADE_TIER_4_TO_TIER_5_CONVERSION.get(),
+            net.p3pp3rf1y.sophisticatedstorage.init.ModItems.BASIC_TIER_UPGRADE.get(),
+            net.p3pp3rf1y.sophisticatedstorage.init.ModItems.BASIC_TO_COPPER_TIER_UPGRADE.get(),
+            net.p3pp3rf1y.sophisticatedstorage.init.ModItems.BASIC_TO_IRON_TIER_UPGRADE.get(),
+            net.p3pp3rf1y.sophisticatedstorage.init.ModItems.BASIC_TO_GOLD_TIER_UPGRADE.get(),
+            net.p3pp3rf1y.sophisticatedstorage.init.ModItems.BASIC_TO_DIAMOND_TIER_UPGRADE.get(),
+            net.p3pp3rf1y.sophisticatedstorage.init.ModItems.BASIC_TO_NETHERITE_TIER_UPGRADE.get(),
+            net.p3pp3rf1y.sophisticatedstorage.init.ModItems.COPPER_TO_IRON_TIER_UPGRADE.get(),
+            net.p3pp3rf1y.sophisticatedstorage.init.ModItems.COPPER_TO_GOLD_TIER_UPGRADE.get(),
+            net.p3pp3rf1y.sophisticatedstorage.init.ModItems.COPPER_TO_DIAMOND_TIER_UPGRADE.get(),
+            net.p3pp3rf1y.sophisticatedstorage.init.ModItems.COPPER_TO_NETHERITE_TIER_UPGRADE.get(),
+            net.p3pp3rf1y.sophisticatedstorage.init.ModItems.IRON_TO_GOLD_TIER_UPGRADE.get(),
+            net.p3pp3rf1y.sophisticatedstorage.init.ModItems.IRON_TO_DIAMOND_TIER_UPGRADE.get(),
+            net.p3pp3rf1y.sophisticatedstorage.init.ModItems.IRON_TO_NETHERITE_TIER_UPGRADE.get(),
+            net.p3pp3rf1y.sophisticatedstorage.init.ModItems.GOLD_TO_DIAMOND_TIER_UPGRADE.get(),
+            net.p3pp3rf1y.sophisticatedstorage.init.ModItems.GOLD_TO_NETHERITE_TIER_UPGRADE.get(),
+            net.p3pp3rf1y.sophisticatedstorage.init.ModItems.DIAMOND_TO_NETHERITE_TIER_UPGRADE.get(),
+            ModBlocks.NETHERITE_BARREL_ITEM.get()
     ));
 }