+++ src/main/java/com/CuteNekoDragon/Core/common/datagen/recipes/Shaped/ShapedRecipesProvider.java
@@ -1,5 +1,6 @@
 package com.CuteNekoDragon.Core.common.datagen.recipes.Shaped;
 
+import com.CuteNekoDragon.Core.common.item.SVOSmithingTemplate;
 import com.gregtechceu.gtceu.data.recipe.VanillaRecipeHelper;
 
 import net.minecraft.advancements.critereon.InventoryChangeTrigger;
@@ -10,7 +11,9 @@ import net.minecraft.tags.ItemTags;
 import net.minecraft.world.item.DyeColor;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.Items;
+import net.minecraft.world.item.crafting.RecipeSerializer;
 import net.p3pp3rf1y.sophisticatedbackpacks.init.ModItems;
+import net.p3pp3rf1y.sophisticatedcore.crafting.ShapeBasedRecipeBuilder;
 import net.p3pp3rf1y.sophisticatedcore.util.ColorHelper;
 
 import com.CuteNekoDragon.Core.SVOCore;
@@ -19,15 +22,25 @@ import com.CuteNekoDragon.Core.common.data.items.SVOItems;
 import com.CuteNekoDragon.Core.common.data.svogt.SVOMachines;
 import com.CuteNekoDragon.Core.common.item.SackItem;
 import com.tterrag.registrate.util.entry.ItemEntry;
+import net.p3pp3rf1y.sophisticatedstorage.crafting.StorageTierUpgradeRecipe;
+import net.p3pp3rf1y.sophisticatedstorage.init.ModBlocks;
 
 import java.util.Map;
 import java.util.function.Consumer;
 
 import static com.CuteNekoDragon.Core.common.data.items.SVOItems.DYED_SACKS;
+import static com.CuteNekoDragon.Core.common.data.items.SVOItems.UPGRADE_TEMPLATES;
 
 public class ShapedRecipesProvider {
 
     public static void buildRecipes(Consumer<FinishedRecipe> consumer) {
+
+        ItemEntry<SVOSmithingTemplate> copperTemplate = UPGRADE_TEMPLATES.get("copper");
+        ItemEntry<SVOSmithingTemplate> ironTemplate = UPGRADE_TEMPLATES.get("iron");
+        ItemEntry<SVOSmithingTemplate> goldTemplate = UPGRADE_TEMPLATES.get("gold");
+        ItemEntry<SVOSmithingTemplate> diamondTemplate = UPGRADE_TEMPLATES.get("diamond");
+        ItemEntry<SVOSmithingTemplate> iridiumTemplate = UPGRADE_TEMPLATES.get("iridium");
+
         ShapedRecipeBuilder.shaped(RecipeCategory.MISC, SVOMachines.CHARKOAL_KILN.getItem())
                 .pattern("ABA")
                 .pattern("A A")
@@ -58,6 +71,44 @@ public class ShapedRecipesProvider {
                 .define('B', SVOItems.SACK)
                 .unlockedBy("has_sack", InventoryChangeTrigger.TriggerInstance.hasItems(SVOItems.SACK))
                 .save(consumer, SVOCore.id("shaped/sack_to_backpack"));
+
+        ShapeBasedRecipeBuilder.shaped(ModBlocks.COPPER_BARREL_ITEM.get(), ModBlocks.STORAGE_TIER_UPGRADE_RECIPE_SERIALIZER.get())
+                .pattern("AAA")
+                .pattern("ABA")
+                .pattern("AAA")
+                .define('A', copperTemplate.get())
+                .define('B', ModBlocks.BARREL_ITEM.get())
+                .unlockedBy("has_barrel", InventoryChangeTrigger.TriggerInstance.hasItems(ModBlocks.BARREL_ITEM.get()))
+                .save(consumer, SVOCore.id("shaped/copper_barrel"));
+
+        ShapeBasedRecipeBuilder.shaped(ModBlocks.IRON_BARREL_ITEM.get(), ModBlocks.STORAGE_TIER_UPGRADE_RECIPE_SERIALIZER.get())
+                .pattern("AAA")
+                .pattern("ABA")
+                .pattern("AAA")
+                .define('A', ironTemplate.get())
+                .define('B', ModBlocks.COPPER_BARREL_ITEM.get())
+                .unlockedBy("has_barrel", InventoryChangeTrigger.TriggerInstance.hasItems(ModBlocks.COPPER_BARREL_ITEM.get()))
+                .save(consumer, SVOCore.id("shaped/iron_barrel"));
+
+        ShapeBasedRecipeBuilder.shaped(ModBlocks.GOLD_BARREL_ITEM.get(), ModBlocks.STORAGE_TIER_UPGRADE_RECIPE_SERIALIZER.get())
+                .pattern("AAA")
+                .pattern("ABA")
+                .pattern("AAA")
+                .define('A', goldTemplate.get())
+                .define('B', ModBlocks.IRON_BARREL_ITEM.get())
+                .unlockedBy("has_barrel", InventoryChangeTrigger.TriggerInstance.hasItems(ModBlocks.GOLD_BARREL_ITEM.get()))
+                .save(consumer, SVOCore.id("shaped/gold_barrel"));
+
+        ShapeBasedRecipeBuilder.shaped(ModBlocks.DIAMOND_BARREL_ITEM.get(), ModBlocks.STORAGE_TIER_UPGRADE_RECIPE_SERIALIZER.get())
+                .pattern("AAA")
+                .pattern("ABA")
+                .pattern("AAA")
+                .define('A', diamondTemplate.get())
+                .define('B', ModBlocks.GOLD_BARREL_ITEM.get())
+                .unlockedBy("has_barrel", InventoryChangeTrigger.TriggerInstance.hasItems(ModBlocks.DIAMOND_BARREL_ITEM.get()))
+                .save(consumer, SVOCore.id("shaped/diamond_barrel"));
+
+
 
 
         for (Map.Entry<DyeColor, ItemEntry<SackItem>> entry : DYED_SACKS.entrySet()) {