+++ changelog.md
@@ -10,6 +10,8 @@
 - customised stack upgrades recipes from sophisticated storage
 - removed stack upgrade from x tier from sophisticated storage to make progression more streamlined
 - customised sophisticated backpacks tier 4 and sophisticated storage tier 5 stack upgrades textures
+- custom sophisticated storage barrel recipes
+- hid netherite barrels
 
 ## [0.0.26]
 ### Changes